Add files via upload
Ultrasonic technology has been on the market for years and is still considered a trusted technology throughout the industry. The design of the Ultrasonic Radar is very useful for many applications like homes, shops, military and object detection. The aim of this work is to build an ultrasonic transceiver which is basically one kind of a radar system to get exact distance and angle for fixed objects placed around the device based on the speed of ultrasonic waves in open air. An Arduino microcontroller was used to transmit and receive the ultrasonic waves through 40 KHz in order to provide the flexibility of usage requirements. A delay occurred between the transmitted and the received waves govern the reflection of sound. Some tests were done using two kinds of alarms first: the visual alarm which done by a personal computer screen designed to be a radar screen. Second the audible beep alarm which done by an LCD digital screen.